Separacion del proyecto en frontend y backend
