+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "vite-project",
+  "name": "frontend",
   "private": true,
   "version": "0.0.0",
   "type": "module",
