+++ frontend/src/components/DetallesCuenta.jsx
