+++ frontend/src/components/FacturaTable.jsx
